--- src/main/java/com/umc/hwaroak/repository/MemberRepository/MemberRepository.java
@@ -1,8 +1,0 @@
-package com.umc.hwaroak.repository.MemberRepository;
-
-import com.umc.hwaroak.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
-
-}
