--- src/main/java/com/umc/hwaroak/repository/MemberRepository/MemberRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.umc.hwaroak.repository.MemberRepository;
-
-public interface MemberRepositoryCustom {
-}
