--- src/main/java/com/umc/hwaroak/repository/MemberRepository/MemberRepositoryImpl.java
@@ -1,9 +1,0 @@
-package com.umc.hwaroak.repository.MemberRepository;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class MemberRepositoryImpl implements MemberRepositoryCustom {
-}
